feat(auth): implement lean JWT token payload
- Reduce token size by including only essential user fields
- Fix ExternalApplicationsModule dependency injection
- Update controllers to use consistent user ID access
