+++ src/external-applications/external-applications.module.ts
@@ -10,13 +10,17 @@ import {
   AdmissionProgram,
   AdmissionProgramSchema,
 } from '../admission-programs/schemas/admission-program.schema';
+import { User, UserSchema } from 'src/users/schemas/user.schema';
+import { UsersModule } from 'src/users/users.module';
 
 @Module({
   imports: [
     MongooseModule.forFeature([
       { name: ExternalApplication.name, schema: ExternalApplicationSchema },
       { name: AdmissionProgram.name, schema: AdmissionProgramSchema },
+      { name: User.name, schema: UserSchema },
     ]),
+    UsersModule,
   ],
   controllers: [ExternalApplicationsController],
   providers: [ExternalApplicationsService],
